Update android_build.yml
Altered the build file to alter the behaviour of the GitHub action.

Previously, the GitHub action that builds the Android app would run on all pull requests into main, regardless of whether the code being pushed requires a rebuild or not.

Following this change, only changes to files in the "Android/" directory will result in a rebuild taking place.

Potential issue: this may be an issue with the branch protection rule that requires status checks to pass. However, this is hard to test without merging this branch.