Added docker-compose.yml
Added a docker compose file to add support for the orchestration  of the SSH Cloud server and the corresponding mysql database server that it accesses.

This involved editing the hibernate configuration file to hard code in the database credentials. _(Hard coding has been chosen as a simplification for the prototyping process.)_

Currently, the docker-compose file only starts the database server - the functionality for it orchestrating the SSH Cloud Docker container is yet to be added.