Debugging database creation and querying
Reduced the number of steps taking place in the Main method to identify where the database creation and querying is failing.

Updated version of Lombok, as it wouldn't build with the previously specified version.

Altered the database schema for debugging purposes.

Also updated the hibernate dialect to a newer version of mysql, as the docker compose file uses the latest available image of mysql for the database server.