Update QueryServicer.java
Added "wrapper functions" to support passing a single object that wraps the parameters to the main functions that service queries.

These are required for implementing the networking, where the sever accepts a request with a single serialised object as a parameter, and thus deserialises it into what is still a single object, that in turn packages the parameters required.

May be refactored at a later point after a functional prototype has been implemented.