Changed the default user
Changed the default hard-coded user in the app so that the user is different to the one that already has items in the shared basket.

This means that the items already in the basket have been added by "Nathan Drake", and the loaded user "Clara Oswald" shouldn't be able to change Nate's items.