Changed how initial database is populated
Removed calls to resources folder for populating the initial database tables. Now, instead of reading in json _files_, the strings from said files have simply been copied into string constants in the new `MoneyBurnerMarketData` class.

The reason for this change is that it was taking too long to debug accessing the files from the resources file after compiling the project into a single .jar file, and for the purposes of this prototype, the fact the data was already hard-coded for simulation, I felt changing _how_ it was hard-coded was somewhat moot.