Update QueryServicer.java
Fixed bug in server method that runs when an item is added to the cart. The bug was that the method didn't check whether it was updating the correct item, and would instead update the first item in the basket. Also fixed the `basketId` being passed into the `increaseItemQuantity` method instead of the (correct) `basketItemId`.