Added a generic client framework
Added a generic client framework that can be used to interact with the SSH Cloud server.

This is the same framework that is used in SSH App.

This has been added to allow for integration tests for SSH Cloud to be written in, and subsequently run automatically in a GitHub Action.